Update README.md
I have a repository in GitHub and I need to tag it.
I tagged in a shell, but on GitHub, it is not showing up.

Do I have to do anything else?

The command I used in the shell is:

git tag 2.0
And now when I type git tag it shows:

2.0
So it seems like tags are present, correct?

The repository is: https://github.com/keevitaja/myseo-pyrocms.

How do I make this tag show up on GitHub? Where are my tags?